requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
requests>=2.31
python-dotenv>=1.0
